gui: Migrate ne2000 PCAP configuration settings
Previously the ne2000 section implied the use of PCAP. This has changed,
but this breaks existing configurations. So migrate them.

The process works like this:
- The ne2000's realnic property is moved to the PCAP section
- If this property was migrated and the ne2000 was enabled, also enable
  the PCAP Ethernet connection

This means in practice people not using the ne2000 code won't be
affected at all, and people who are will have their PCAP-based
networking continue working as usual.