ethernet_slirp: Initial Windows support
This code currently works in Wine on Linux, but seems to have a bug on
ReactOS due to MSG_PEEK eating the input. More testing is needed but so
far this seems to work when compiled with MinGW.

Windows XP support requires small changes to libslirp to remove its use
of inet_ntop and inet_pton functions. These are only used for IPv6 and
Linux /etc/resolv.conf parsing and aren't needed at the moment.